Prevent Active Record instances from being used as params in test requests
This upstreams a runtime check we have to prevent invalid params in test requests that could lead to false positives in testing. We check for a variety of things, some of which are domain specific. Something that is *not* domain specific is Active Record instances - there's no scenario I can think of where `params[:key]` should be an AR instance, as opposed to its ID.

With this PR, tests that have incorrect params will raise an `ActionDispatch::IntegrationTest::InvalidParamError`.