Refactor offsets method to avoid race condition
In environments where the routes aren't eager loaded such as development or
test a race condition could occur where two requests would trigger the
`offsets` method in `ActionDispatch::Journey::Path::Pattern` and the early
return for `@offsets` would be triggered by the initialisation of the
instance variable to `[0]` before the rest of the method had completed.

Fixes #43431.