Group `validates_associated` documentation with `validates_*` [ci-skip]
`validates_associated` makes sure that passed associations are validated
everytime the record is saved. This is unlike other validations that
validate an attribute/association has a certain value.
Typically you won't call:

    validates :books, associated: true

By grouping `validates_associated` with `validates_with` and
`validates_each` it's made clearer `validates_associated` is a bit
different.

With this change the validations are sorted alphabetically.