Add a way to limit unbounded IN queries
__NOTE: This is just an experiment. Please DO NOT merge.__

It's currently really easy to create unbounded `IN` queries that grow
large without the application developer accounting appropriately for
the size.

This commit introduces a setting which requires the application
developer to state up front the maximum size of a given `IN`.

When this setting is on, rather than a traditional `where`, they'll need
to specify the maximum size of the `IN` portion:

``` ruby
Article.where(id: selected_ids.bound_at(5))
```

With this, if the amount of `ids` is more than `5`, the application
would only send the first `5` forward to the `IN` query.

Fundamentally, the idea here is about making application developers
aware of the queries that they are making rather than making it very
easy to introduce these unbounded `IN`s.

Feedback more than welcome.