forgetting_assignment avoid serialization
`#forgetting_assignment` was introduced with 07723c23a7d, and it
involved deserializing and then serializing current value.

This is an unnecessary extra operation, and also it
is a problem for custom attribute types where `#serialize` and
`#deserialize` are not inverse to each other operations.

This is a problem for example when using oracle-enhanced-adapter.

fixes #44317 and #42738 and rsim/oracle-enhanced#2268