Add info for `change` column and table comment
This will ensure the guide is aligned with the doc on some things that I believe it's useful. Some folks do go for the guides more of than the api for a quick reference.

Order alphabetically

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>

Add change colum and table comment links for reference in guide

Update  working so we know this list may keep changing