Expose request method to reset the CSRF token
We already have a commit CSRF method exposed via the request object
since it's used by the implementation when committing the session, so
having a similar reset CSRF method exposed makes sense, and hides some
of the internal complexity of calling that method via the controller
instance.

It will also facilitate reaching out to the reset CSRF logic from other
libraries like Devise, to more easily integrate with this change.