Only materialize the new transaction if it needs it
Historically this was unnecessary, because all possible Transaction
instances would start their life unmaterialized. But now, if the new
transaction is a ResetParentTransaction, its materialization state
relies upon its parent, so it may already be done. Indeed, if lazy
transactions are disabled, the parent is almost certainly already
materialized. #materialize! will happily re-run the query without
question (a behaviour we rely upon elsewhere), so in this case it's up
to the caller to check first.