Introduce `config/application/` directory
Historically, configs were supposed to be entirely set from `application.rb`
and the various `config/environments/` files.

But Rails itself broke this assumption by setting some configurations
from initializer files. Actually all the default initializer files
are setting configurations.

The problem with this is that it becomes very hard to know when a specific
configuration is considered final and is applied. So this had let to delay
applying some configurations so that they can be defined in initializers.

But it's very easy to set a configuration too late and not realize it's not
applied.

The idea here is two fold.

First we introduce `config/application/` so that we better support the pattern
of breaking down `application.rb` in smaller files.

Then we can warn or error out when the `config` object is mutated past
the point where it no longer is applied.