Deprecate plural association names on singular associations
Before this change associations could be called by their plural or singular name,

i.e.

	  class Post < ActiveRecord::Base
	    has_many :comments
	  end

	  class Comment < ActiveRecord::Base
	    belongs_to :post
	  end

	  # this works with `post` as the key:
	  Comment.where(post: Post.first)
	  # this also works with `:posts` as the key (after this change this will warn):
	  Comment.where(posts: Post.first)

This PR deprecates using the plural association name when the association was not defined as plural.

Motivated by https://github.com/rails/rails/pull/45020 which found that singularize call is slow and happens very often when you are querying a column name which is not a relation, so it must try both paths. After this deprecation cycle we can save that time in the `where(column_name: "value")` case.

After this deprecation cycle, we can also avoid a potentially annoying edge case where you are unable to query a column with a singular name the same as your association's name.

It seems this singularize was added 2 years ago here: https://github.com/rails/rails/pull/39465 , however inheritance and calculation tests still with this change removed.