Bubble assertion failures back to the main thread
Without some help, failures in a forked process make for some noise in
the output, but won't fail the build.

Instead of trying to transfer the whole exception back, I've gone for a
simpler solution of just sending _something_ (the exception class name)
back so we'll fail; the full failure will be visible in the child
process's stderr output.