Support unbounded time ranges for PostgreSQL
Ruby raises an `ArgumentError` when mixing `Time` and `Float` values in
a `Range`, such as `Time.now..Float::INFINITY`.  Therefore, when
deserializing time ranges, we must use a beginless range to represent a
range with no lower bound, or an endless range to represent a range with
no upper bound.

Fixes #39833.
Closes #45082.

Co-authored-by: fatkodima <fatkodima123@gmail.com>