Allow setting db with use_postgresql tests
If more than one test is using `use_postgresql` for it's tests and both
need to run `db:create` those tests will fail often due to them
running in parallel. I tried turning off parallelization for these tests
and they still ran in separate processes. Ultimately though, turning off
parallelization means we don't aren't testing parallelization. Instead I've
added `Process.pid` to the database name so it creates a database
specifically for that test. For cases where we need to set an explict
name a name can be passed in.

I also added an ensure to one of the tests so that the databases get
cleaned up and not left behind.

Fixes #45114
Fixes #45158

Manual backport of #45178