Application-specific `rails routes` option
By building the routes and checking where the routes are defined at `.draw` time
we can determine whether the current block of routes are those defined in the
users application, or in services like active_storage, action_mailbox etc.

When running `rails routes --application` we will render only routes defined in
the application (routes.rb, engines included in routes.rb etc)

Limiting the routes to just those defined by the user, it allows for easier
reading and a clearer view of routes.