Temporarily remove platforms from tzinfo-data gem
The current platforms string will not match for Ruby 3.1+ on Windows,
causing tzdata-info to not be installed and loading the environment
during `rails new` to fail.

For now, we should remove the platforms restriction so that all new
application Gemfiles will include tzdata-info by default. This provides
a better experience for those running Ruby 3.1 on Windows, and if the
user knows they will not be running on Windows then they can remove the
gem or apply the platforms optimization themselves.

This change should be reverted (or updated) after rubygems/rubygems#5269
is resolved.