Allows associations to be automatically preloadable. Before we load an association for the first time, we can preload the association. This will eliminate most standard N+1 queries in Rails apps.
We do this by capturing the group of records returned by a query in an
AutomaticPreloader and having each record have a reference to it.

Once a record has been flagged to allow automatic preloading, subsequent
associations are also flagged the same so this will allow deeply nested
associations to all be dynamically preloaded

This is off by default and can be opted in with each relation. There
is also a global option to enable it by default