feat: Add Colorized, a new routes formatter
I added `ActionDispatch::Routing::ConsoleFormatter::Colorized` that is a new routes formatter.

Why a new formatter?

The intention  of this one is to add a different structure and colors to the actual `rails route` command.
This formatter can be used by a new flag (`-C`) added on the `route` command.