Show Guides Index as selected for `/` on small screens [ci-skip]
On small screens the guides index is shown as a select dropdown.
The Guides Index option has the value 'index.html'.
When visiting `https://guides.rubyonrails.org/` this option won't be
selected as the pathname doesn't include index.html.

Javascript treats empty strings as falsey, so for the root path we can
return `index.html` instead of a empty string.