Introduce ActiveSupport::ConfigurationOptions
Ref: https://github.com/rails/rails/pull/44980
Ref: https://github.com/rails/rails/pull/44996

A very common problem we're running into are configurations
that are set after they are applied. It's particularly bad
because it silently have no effect.

If initializers are explicit when they consume a configuration
value, it would allow use to raise when a user changes it later.

Instead of just accessing the config, we call `consume(:my_config)`
and any later attempt at changing it will fail.

Of course this requires to rewrite most initializers, and 3rd party
railties will likely lag behind for a while, but even then we can
effectively solve the problem for all 1st party configurations.