Enable testing of binary data in request specs
Previously it was not possible to test POST requests with binary data
with no `Content-Type` header. rack-test will default to a
`Content-Type` of `application/x-www-form-urlencoded`, which causes
encoding errors if binary data is passed in the request parameters.

If an `:input` parameter is passed to `Rack::Test::Session`, then
`Rack::Test::Session` will not attempt to set the `Content-Type`, and
the raw data will be passed to the body of the POST request. We can
use this to test binary data with no `Content-Type`.

This came up in https://github.com/rack/rack-test/issues/247.