Reset cache_versions on relation
Currently, when #reset is called on a relation object it does not reset
the cache_versions ivar. This can lead to a confusing and buggy
situation where despite having the correct data the relation is still
reporting a stale cache_version. Resetting this ivar along with the
other relation state corrects this situation.

Update test assertion

Assert the specific cache_version matches the latest .all relation

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>