Update Command Line guide --database usage [ci-skip]
The current Command Line guide has an advanced section describing
specifying git as SCM and PostgreSQL as the database. This section is
easily overlooked: https://github.com/rails/rails/issues/44325

By moving whole "advanced" section under the `rails new` section, it's
easier to find. Also specifying the database is pretty common I guess
and not an "advanced" topic.

All mentions of the SCM option have been removed, as it seems to no
longer be an option and we always initialize with Git only.