Move Intranet Security guide [ci-skip]
The Intranet Security section feels a bit out of place between the
sections describing issues that apply to all applications: CSRF,
Redirection and Files, User Management and Injection sections.
It also mentions Cross-Site Scripting which is described in detail later
on.

As it's more of a meta section, referencing previous sections, it is
moved above the Environmental Security guide. The XSS and CSRF section
have also been given proper headings to make the section more easy to
scan.