Scaffold destroy action returns status code 303
The default controller scaffolding currently sends back a 302 for its
destroy action.

Since Hotwire is the default for Rails going forward, this can be
problimatic. Turbo uses the fetch API internally which is particular on
how it handles redirects.

As outlined in this table in this Turbo issue,
https://github.com/hotwired/turbo/issues/84#issuecomment-862656931,
Turbo making a DELETE request (not POST + hidden _method field) and
recieving a 302 back will result in another DELETE request instead of a
GET request for the redirect.

This updates the controller template used for the scaffold generator
to send back the 303 see other status code for the destroy action.

For consistancy-sake, the Action Controller Overview guide examples
were also updated. The main Getting Started guide page already has
see other used in its example destroy action.

For non-Hotwire users, the browser will still properly redirect on
a 303 as it does with a 302.