Add separate guides page listing Rails versions
The purpose of this page is to provide a single location that any
version of the guides can link to in order for the user to navigate to
any other version of the guides.

For example, https://guides.rubyonrails.org/v6.1/index.html can link to
https://guides.rubyonrails.org/versions.html, and the user will be able
to navigate to https://guides.rubyonrails.org/v7.0/index.html, even
after v7.0 is no longer the latest release and the v6.1 guides are no
longer updated.

This commit also links the version badge displayed at the top of each
page to the versions page.